refactor: Move update.js from ui/ to lib/
