+++ lib/helpers/update$dteTimezone.js
@@ -1,5 +1,5 @@
 import moment from 'moment-timezone';
-import { List, Map, OrderedMap, Set } from 'immutable';
+import { List, Map, OrderedMap, Set, fromJS } from 'immutable';
 
 /**
  * Change $dte timezone offset
@@ -15,7 +15,7 @@ import { List, Map, OrderedMap, Set } from 'immutable';
  * @param {string} timezone - TZ database name e.g. "Europe/Paris"
  * @returns {any}
  */
-const update$dteTimezone = (query, timezone) => {
+export const update$dteTimezone = (query, timezone) => {
   // if query has mapEntries()
   if (Map.isMap(query) || OrderedMap.isOrderedMap(query)) {
     return query.mapEntries(([k, v]) => {
@@ -38,4 +38,15 @@ const update$dteTimezone = (query, timezone) => {
   return query;
 };
 
-export default update$dteTimezone;
+/**
+ * String -> JSON -> Immutable -> {update} -> Immutable -> JSON -> String
+ *
+ * @param {string} queryString
+ * @param {string} timezone - TZ database name e.g. "Europe/Paris"
+ * @returns {string}
+ */
+export const update$dteTimezoneString = (queryString, timezone) => {
+  const query = fromJS(JSON.parse(queryString));
+  const updated = update$dteTimezone(query, timezone);
+  return JSON.stringify(updated.toJS());
+};
