+++ ui/src/components/StatementForwarding/StatementForwardingForm.js
@@ -4,12 +4,12 @@ import { List, Map } from 'immutable';
 import classNames from 'classnames';
 import { withProps, compose, withHandlers } from 'recompose';
 import { getAuthHeaders, getHeaders } from 'lib/helpers/statementForwarding';
+import { update$dteTimezone } from 'lib/helpers/update$dteTimezone';
 import { AUTH_TYPES, AUTH_TYPE_NO_AUTH, STATAMENT_FORWARDING_MAX_RETRIES } from 'lib/constants/statementForwarding';
 import Switch from 'ui/components/Material/Switch';
 import ValidationList from 'ui/components/ValidationList';
 import TableInput from 'ui/components/TableInput';
 import { TimezoneSelector, buildDefaultOptionLabel } from 'ui/components/TimezoneSelector';
-import update$dteTimezone from 'ui/utils/queries/update$dteTimezone';
 import QueryBuilder from 'ui/containers/QueryBuilder';
 import activeOrgSelector from 'ui/redux/modules/activeOrgSelector';
 import { withModel } from 'ui/utils/hocs';
